♻️ Refactor: 모임 정산 내역을 공유 기능 로직을 services/receiptService.ts로 이동
