+++ src/App.tsx
@@ -15,6 +15,7 @@ import InputPaymentHistory from './components/InputPaymentHistory';
 import InputReceiptInfo from './components/InputReceiptInfo';
 import { useReceipt } from './hooks/useReceipt';
 import PaymentHistoryList from './components/PaymentHistory';
+import { createReceipt } from './services/receiptService';
 
 function App() {
   const {
@@ -29,55 +30,9 @@ function App() {
   const [isShareOpen, setIsShareOpen] = useState(false);
   const [isCopied, setIsCopied] = useState(false);
 
-  const getDisplayLength = (str: string) => {
-    let length = 0;
-    for (let i = 0; i < str.length; i += 1) {
-      const char = str.charAt(i);
-      // 한글은 2자리, 영문/숫자는 1자리
-      if (char.match(/[ㄱ-ㅎ|ㅏ-ㅣ|가-힣]/)) {
-        length += 2;
-      } else {
-        length += 1;
-      }
-    }
-    return length;
-  };
-
-  const createReceipt = () => {
-    const WIDTH = 25;
-    const LINE = '='.repeat(WIDTH);
-
-    const title = `[💳 ${receipt.title}]\n`;
-    const items = receipt.histories.map((history) => {
-      const priceText = `${history.amount.toLocaleString()}원`;
-      const contentLen = getDisplayLength(history.content);
-      const priceLen = getDisplayLength(priceText);
-
-      const spaces =
-        contentLen + priceLen < WIDTH ? WIDTH - contentLen - priceLen : 0;
-
-      return `${history.content}: ${' '.repeat(spaces)}${priceText}`;
-    });
-    const totalPrice = receipt.histories.reduce(
-      (acc, curr) => acc + curr.amount,
-      0,
-    );
-    const totalPriceStr = `총 금액: ${' '.repeat(WIDTH - getDisplayLength('총 금액: ') - getDisplayLength(`${totalPrice.toLocaleString()}원`))}${totalPrice.toLocaleString()}원`;
-
-    const pricePerPerson = `1/N 금액: ${' '.repeat(
-      WIDTH -
-        getDisplayLength('1/N 금액: ') -
-        getDisplayLength(
-          `${Math.ceil(totalPrice / receipt.peopleCount).toLocaleString()}원`,
-        ),
-    )}${Math.ceil(totalPrice / receipt.peopleCount).toLocaleString()}원`;
-
-    return [title, ...items, LINE, totalPriceStr, pricePerPerson].join('\n');
-  };
-
   const handleCopyButton = async () => {
     try {
-      const created = createReceipt();
+      const created = createReceipt(receipt);
 
       await navigator.clipboard.writeText(created);
       setIsCopied(true);
